+++ examples/schema_create.py
@@ -1,6 +1,7 @@
 import asyncio
 
 import datetime
+import secrets
 
 from examples import get_db_name
 from tortoise import Tortoise, fields
@@ -11,20 +12,25 @@ from tortoise.utils import generate_schema
 
 class Tournament(Model):
     id = fields.IntField(pk=True)
-    name = fields.StringField()
+    name = fields.TextField()
     created = fields.DatetimeField(auto_now_add=True)
 
     def __str__(self):
         return self.name
 
 
+def generate_token():
+    return secrets.token_hex(16)
+
+
 class Event(Model):
     id = fields.IntField(pk=True)
-    name = fields.StringField(unique=True)
+    name = fields.TextField(unique=True)
     tournament = fields.ForeignKeyField('models.Tournament', related_name='events')
     participants = fields.ManyToManyField('models.Team', related_name='events', through='event_team')
     modified = fields.DatetimeField(auto_now=True)
     prize = fields.DecimalField(max_digits=10, decimal_places=2, null=True)
+    token = fields.TextField(default=generate_token)
 
     def __str__(self):
         return self.name
@@ -53,6 +59,7 @@ async def run():
     event.name = 'Test 2'
     await event.save()
     assert event.modified > old_time
+    assert len(event.token) == 32
 
 
 if __name__ == '__main__':