+++ examples/transactions.py
@@ -18,7 +18,7 @@ you could still do it like this
 
 class Event(Model):
     id = fields.IntField(pk=True)
-    name = fields.StringField()
+    name = fields.TextField()
 
     class Meta:
         table = 'event'