+++ examples/two_databases.py
@@ -21,7 +21,7 @@ and explicitly declaring model apps in class Meta
 
 class Tournament(Model):
     id = fields.IntField(pk=True)
-    name = fields.StringField()
+    name = fields.TextField()
 
     def __str__(self):
         return self.name
@@ -32,7 +32,7 @@ class Tournament(Model):
 
 class Event(Model):
     id = fields.IntField(pk=True)
-    name = fields.StringField()
+    name = fields.TextField()
     tournament_id = fields.IntField()
     # Here we make link to events.Team, not models.Team
     participants = fields.ManyToManyField('events.Team', related_name='events', through='event_team')
@@ -46,7 +46,7 @@ class Event(Model):
 
 class Team(Model):
     id = fields.IntField(pk=True)
-    name = fields.StringField()
+    name = fields.TextField()
 
     def __str__(self):
         return self.name